refactoring sign_as
