+++ src/commands/account/create_account/fund_myself_create_account/add_key/autogenerate_new_keypair/mod.rs
@@ -149,7 +149,7 @@ impl SaveMode {
 pub struct SignAs {
     #[interactive_clap(named_arg)]
     ///What is the signer account ID?
-    sign_as: super::super::SignerAccountId,
+    sign_as: super::super::sign_as::SignerAccountId,
 }
 
 impl SignAs {
