+++ src/commands/account/create_account/fund_myself_create_account/add_key/use_manually_provided_seed_phrase/mod.rs
@@ -7,7 +7,7 @@ pub struct AddAccessWithSeedPhraseAction {
     master_seed_phrase: String,
     #[interactive_clap(named_arg)]
     ///What is the signer account ID?
-    sign_as: super::super::SignerAccountId,
+    sign_as: super::super::sign_as::SignerAccountId,
 }
 
 impl AddAccessWithSeedPhraseAction {
